Move `Analytics` to the `:simple-platform` module
