+++ app/src/main/java/org/simple/clinic/ReportAnalyticsEvents.kt
@@ -3,7 +3,7 @@ package org.simple.clinic
 import io.reactivex.Observable
 import io.reactivex.ObservableSource
 import io.reactivex.ObservableTransformer
-import org.simple.clinic.analytics.Analytics
+import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.widgets.UiEvent
 
 class ReportAnalyticsEvents : ObservableTransformer<UiEvent, UiEvent> {
