+++ app/src/main/java/org/simple/clinic/analytics/ErrorSwallowingReporter.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.analytics
 
+import org.simple.clinic.platform.analytics.AnalyticsReporter
 import org.simple.clinic.platform.analytics.AnalyticsUser
 import timber.log.Timber
