+++ app/src/main/java/org/simple/clinic/analytics/UpdateAnalyticsUserId.kt
@@ -1,6 +1,7 @@
 package org.simple.clinic.analytics
 
 import android.annotation.SuppressLint
+import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.platform.analytics.AnalyticsUser
 import org.simple.clinic.user.User.LoggedInStatus.LOGGED_IN
 import org.simple.clinic.user.User.LoggedInStatus.RESETTING_PIN
