+++ app/src/main/java/org/simple/clinic/user/registeruser/RegisterUser.kt
@@ -3,7 +3,7 @@ package org.simple.clinic.user.registeruser
 import com.f2prateek.rx.preferences2.Preference
 import io.reactivex.Completable
 import io.reactivex.Single
-import org.simple.clinic.analytics.Analytics
+import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.login.UsersApi
