+++ app/src/test/java/org/simple/clinic/ReportAnalyticsEventsTest.kt
@@ -6,7 +6,7 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.simple.clinic.analytics.Analytics
+import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.analytics.MockAnalyticsReporter
 import org.simple.clinic.analytics.MockAnalyticsReporter.Event
 import org.simple.clinic.util.RxErrorsRule
