+++ app/src/test/java/org/simple/clinic/analytics/ErrorSwallowingReporterTest.kt
@@ -2,6 +2,7 @@ package org.simple.clinic.analytics
 
 import org.junit.Test
 import org.simple.clinic.TestData
+import org.simple.clinic.platform.analytics.AnalyticsReporter
 import org.simple.clinic.platform.analytics.AnalyticsUser
 
 class ErrorSwallowingReporterTest {
