+++ app/src/test/java/org/simple/clinic/analytics/RxTimingAnalyticsTest.kt
@@ -6,6 +6,7 @@ import io.reactivex.subjects.PublishSubject
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
+import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.util.advanceTimeBy
 import org.threeten.bp.Duration
