+++ app/src/test/java/org/simple/clinic/user/ReportPendingRecordsToAnalyticsTest.kt
@@ -10,7 +10,7 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.simple.clinic.analytics.Analytics
+import org.simple.clinic.platform.analytics.Analytics
 import org.simple.clinic.analytics.MockAnalyticsReporter
 import org.simple.clinic.analytics.MockAnalyticsReporter.Event
 import org.simple.clinic.bp.BloodPressureMeasurement
